requirements.txt
# SUMO Traffic Simulation
traci
sumolib
eclipse-sumo

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
imbalanced-learn

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.25.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
